align: extract 9 medium inline test blocks (rule 8 part 2)
Continues the rule-8 sweep from the previous commit. Brings the
inline-tests-extracted total from 7/22 to 16/22; the remaining
6 are >200 lines each and queued as a separate large-block PR per
the updated plan.

Files moved (medium tier, 47-194 inline lines each):
- orders/builder/condition_helpers.rs -> condition_helpers_tests.rs
- trace/sync.rs                       -> sync_tests.rs
- trace/async.rs                      -> async_tests.rs
- client/id_generator.rs              -> id_generator_tests.rs
- wsh/common/decoders.rs              -> decoders_tests.rs
- orders/conditions.rs                -> conditions_tests.rs
- transport/recorder.rs               -> recorder_tests.rs
- common/error_helpers.rs             -> error_helpers_tests.rs
- common/retry.rs                     -> retry_tests.rs

common/retry preserves the inner
`#[cfg(feature = "sync")] mod sync_tests {...}` /
`#[cfg(feature = "async")] mod async_tests {...}` substructure
unchanged - just moves the outer wrapper.

plans/code-consistency-followups.md updated: small + medium batches
now listed under done; the deferred table is trimmed to the 6 large
files that remain.

Gates green on this commit: cargo fmt; cargo clippy --all-targets
under default / sync / all-features (each -D warnings);
cargo test --features sync (244 passed, 0 failed).
